fix: 14项缺陷修复（安全/并发/Git状态/信号/阻塞IO）

+++ src/config.rs
@@ -153,7 +153,7 @@ impl AppConfig {
 
         let source = ScanSource {
             root_path: path_str,
-            max_depth: self.default_depth as i32,
+            max_depth: self.default_depth,
             ignore_patterns: self.ignore_patterns.clone(),
             ..Default::default()
         };
